fix2

+++ modules/task_1/eremin_a_vector_sum/ops_mpi.cpp
@@ -42,7 +42,7 @@ int sumParallel(std::vector<int> Vector, int size) {
     std::vector<int> local_vec(delta);
     if (ProcRank == 0) {
         local_vec = std::vector<int>(Vector.begin(), Vector.begin() + delta);
-        for (int i = 1; i < tmp; i++) {
+        for (size_t i = 1; i < tmp; i++) {
             MPI_Send(Vector.data() + i * delta, delta, MPI_INT, i, 0, MPI_COMM_WORLD);
         }
     } else if (ProcRank < size) {
@@ -51,7 +51,7 @@ int sumParallel(std::vector<int> Vector, int size) {
     }
     if (ProcRank == 0) {
         if (size % tmp > 0) {
-            int j = 0;
+            size_t j = 0;
             local_vec.resize(local_vec.size() + size % ProcNum);
             for (int i = (size - size % ProcNum); i < size; i++) {
                 local_vec[j + delta] = Vector[i];
